Create database connection
When working with a database a connection is required to perform any 
queries or operations. The connection is tied to a request and is closed 
before the response is sent.

**g** is a special object that is unique for each request and is used to 
store data during the request. The connection is stored on the key 'db'.

**current_app** is a special object that points to the Flask app 
handling the request. This is required as an application factory is used 
and there is no application object when writing the code.

The 'DATABASE' configuration key points to the database file. To return 
rows that behave like dicts sqlite3.Row is passed to the row_factory.

The close_db function checks if a connection was created, and closes it 
if it exists.