Create CLI command to execute SQL schema
A function is required to run the SQL commands that will create the new 
user and post tables. Use a context manager open the schema file and 
execute the commands against the connected database. The file is opened 
relative to the flaskr package.

We will use the click module to create a CLI command that calls the new 
function and displays a success message.