Create authentication blueprint
A flask **Blueprint** called 'auth' is required. Blueprints are used to 
organize groups of related views and code. The authentication blueprint 
will have views to register new users, log in and log out.

The blueprint will be called 'auth' and the prefix '/auth' will be 
prepended to all associated URLs.