Create register view
When the user visits the `/auth/register` URL, the register view should
return a HTML form where they can submit a username and password.

The route decorator associates the url with the view function.

If the request object has the method set to 'GET', the register template
will be rendered.

If the method is set to 'POST', the username and password will be taken
from the form dict which is a special type mapping submitting form keys
and values. These will be inserted into the user table, with the
password hashed. The changes are committed and the user is redirected to
the login page.