Add register validation to server
While the register template has a required attribute for the username 
and password, the user may be using a browser that does not support it, 
or may be making a request using something other than a browser. This 
means username and password validation should also be added to the 
server.

Use the flash module, which stores messages that can be retrieved when 
rendering the template.