Add username not unique validation
Register validation should also check if the username is unique. Add a 
check that executes a SQL statement on the database, returning an id for 
the username that has been entered.

The database library takes care of escaping values so that it is not 
vulnerable to a SQL injection attack. `fetchone` will return only one 
row from the query.