Create login view
When the user visits the `/auth/login` URL, the login view should return 
a HTML form where they can submit a username and password.

When a subsequent post request is made the username is queried in the 
database, with results assigned to a variable.

If user data is returned, the password is checked by hashing it in the 
same way as the stored hash and comparing them.

The `user_id` is finally stored in a new session dictionary which is 
saved across requests. This means the `user_id` will be available on 
subsequent requests.