Add load logged in user function
At the beginning of each request, if a user is logged in their 
information should be loaded and made available to other views.

Use the `before_app_request` blueprint function to execute a 
`load_logged_is_user` function that will execute before each request, 
even if outside of the blueprint.

The function will use attempt to use the `user_id` stored in the session 
to populate the `g.user` value if the user has already logged in.