Add dynamic header to base template
The information and navigation links in the top banner should 
dynamically change depending on whether the user has logged in.

The g object can be used to assert if a user has logged in, to then 
either show their username and a logout button if they have. Or a 
register and login button if they have not logged in.