Add post method to blog create view
The post method should use the title and body from the request
submission, along with the user id which can be take from the user dict
held on the g object.

The g object will be populated as the user should be logged in when
creating a post, and the method `load_logged_in_user` is executed before
each request.