Add login_required decorator
Creating a blog post will require a user to be logged in. Use a
decorator to check this for each view it is applied too.

The decorator returns a new view function that wraps the original view
it is applied too. The new function checks if a user is loaded and
redirects to the login page otherwise. It the user is populated the
original view and any arguements that were passed to it is called.