Create blog index view
The index view for the blog should show all of the posts, with the most
recent first. The data query includes a join between the posts and user
table so that the author can be identified.

Post data is assigned to a variable and passed to the index template via
a second arguement.