Create blog index template
The index page should loop through each post, displaying the title,
username, created date and body.

Each post will be seperated by a horizontal line `hr`. The `loop.last`
variable is specially available inside Jinja for loops, and is used here
so that the line is not added after the last post.