Add index url rule
The endpoint for the index view `/` of the blog is currently defined as
`blog.index`. As this will be the main index for the site we also need
to associate the endpoint name `index` with `/`.

Using `app.add_url_rule()` both `url_for('index')` and
`url_for('blog.index')` will generate the same `/` url.