Create blog update view
The update view should fetch a `post` by `id` and allow the user to
modify the title and/or body before saving the updated post.

The update view requires an arguement `id` that corresponds to the `id`
of the `post` that is being requested to update. This is specified in
the route decorator, with `<int:id>` capturing the number from the url,
ensuring it is an `int`, and passing it as an arguement to the function.

Return an update template, passing the post details assigned to the
`post` variable as an arguement.