Add post update exceptions
If the post the user is attempting to update does not exist, or they are
trying to update a post that they have not created, exceptions should be
raised.

Use `abort()` to raise a special exception that returns a HTTP status
code. 404 is used when a page is "Not Found", and 403 means "Forbidden".