Create get_post helper function
Both the blog update and delete views need to fetch a post by id and
check if the author matches the logged in user. To avoid duplication
create a helper function and call it from each view.

The helper function should retrieve the post details from the database
using the id, or return an error based on whether the post exists, or if
the user is not the author.