Create project setup to make installable
So that the project can be installed in other environments create a
distribution file that describes the project and the files that belong
to it.

This makes deploying the project the same as installing any other
library.

`packages` tells Python what package directories and the python files
they contain to include. The `find_packages()` function finds these
directories automatically. To include static and template dictories,
`include_package_data` is set.